Consolidate views, add sprockets, configure JS/Fonts/CSS
