+++ .gitignore
@@ -34,3 +34,9 @@
 
 # Ignore coverage directory
 /coverage/*
+/public/packs
+/public/packs-test
+/node_modules
+/yarn-error.log
+yarn-debug.log*
+.yarn-integrity
