+++ app/assets/config/manifest.js
@@ -1,2 +1,4 @@
 //= link_tree ../images
 //= link_directory ../stylesheets .css
+//= link_directory ../javascript .js
+//= link_directory ../fonts
