+++ app/controllers/saasbook_controller.rb
@@ -14,14 +14,15 @@ class SaasbookController < ApplicationController
     @chapter_id = params[:chapter_id]
     @section_id = params[:section_id]
 
-    # This will eventually be replaced with a database + model construction
-    render("chapter#{@chapter_id}section#{@section_id}")
+    @body_contents = "chapter#{@chapter_id}section#{@section_id}"
+
+    render("book_content")
   end
 
   def show_chapter
     @chapter_id = params[:chapter_id]
+    @body_contents = "chapter#{@chapter_id}"
 
-    # This will eventually be replaced with a database + model construction
-    render("chapter#{@chapter_id}")
+    render("book_content")
   end
 end
