+++ app/javascript/packs/saas.js
@@ -1,0 +1,2 @@
+import 'src/sphinx-book-theme.js'
+import 'src/doctools.js'
