+++ app/javascript/src/jquery.js
