+++ config/application.rb
@@ -27,7 +27,6 @@ module Esaasbook
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.1
-
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
@@ -38,5 +37,6 @@ module Esaasbook
 
     # Don't generate system test files.
     config.generators.system_tests = nil
+    config.assets.paths << Rails.root.join("app", "assets", "fonts")
   end
 end
