+++ package.json
@@ -2,9 +2,17 @@
   "name": "esaasbook",
   "private": true,
   "dependencies": {
-    "@rails/ujs": "^6.0.0",
+    "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
-    "@rails/actioncable": "^6.0.0"
+    "@rails/ujs": "^6.0.0",
+    "@rails/webpacker": "5.4.3",
+    "jquery": "^3.6.0",
+    "underscore": "^1.13.2",
+    "webpack": "^4.46.0",
+    "webpack-cli": "^3.3.12"
   },
-  "version": "0.1.0"
+  "version": "0.1.0",
+  "devDependencies": {
+    "webpack-dev-server": "^3"
+  }
 }
